LayoutTests/fullscreen/full-screen-enabled.html

<body>
<div>This tests the <code>fullscreenEnabled</code> property laid out in section 4 of the W3C 
<a href="http://dvcs.w3.org/hg/fullscreen/raw-file/tip/Overview.html">Full Screen API</a></div>
<script src="full-screen-test.js"></script>
<script>
    var iframe = document.createElement('iframe');
    iframe.setAttribute('allowfullscreen', 'true');
    document.documentElement.appendChild(iframe);
    var iframe2 = document.createElement('iframe')
    iframe2.setAttribute('allow', "fullscreen 'none'");
    iframe2 = document.documentElement.appendChild(iframe2);
    testExpected('iframe.contentDocument.webkitFullscreenEnabled', true);
    testExpected('iframe2.contentDocument.webkitFullscreenEnabled', false);
    endTest();
</script>
